requirements.txt
langchain-openai>=1.1.0
langchain-core>=1.1.0
langchain>=1.1.0
langgraph>=1.0.10
langchain-azure-ai>=1.0.4
langchain-mcp-adapters>=0.1.14
python-dotenv>=1.2.1
sqlmodel>=0.0.22
psycopg[binary]>=3.1.18
pgvector>=0.2.5
rich>=13.7.0
questionary>=2.0.1
